requirements.txt
absl-py==2.3.1
accelerate==1.10.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
bitsandbytes==0.47.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
datasets==4.0.0
dill==0.3.8
evaluate==0.4.5
filelock==3.13.1
frozenlist==1.7.0
fsspec==2024.6.1
grpcio==1.74.0
hf-xet==1.1.7
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.4
joblib==1.5.1
Markdown==3.8.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.3
nltk==3.9.1
numpy==2.1.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.3.1
peft==0.17.0
pillow==11.0.0
propcache==0.3.2
protobuf==3.20.1
psutil==7.0.0
pyarrow==21.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rouge_score==0.1.2
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==70.2.0
simpleitk==2.5.2
six==1.17.0
sympy==1.13.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.21.4
tqdm==4.67.1
transformers==4.55.2
triton==3.4.0
trl==0.21.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1

--extra-index-url https://download.pytorch.org/whl/cu126
torch==2.8.0+cu126
torchaudio==2.8.0+cu126
torchvision==0.23.0+cu126